fix predicate definition init for mona, rename mona dir
